pyproject.toml
[tool.poetry]
name = "autoshorts"
version = "0.1.1"
description = ""
authors = ["Roman Pisani <romanp2929@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
moviepy = "^1.0.3"
openai = "^1.35.10"
numpy = "^2.0.0"
pydub = "^0.25.1"
whisper = "^1.1.10"
requests = "^2.32.3"
asyncio = "^3.4.3"
nltk = "^3.8.1"
google-generativeai = "^0.8.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"